requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn

pymongo[srv]==3.6
# used for when tou want to develop some program into pypi package
## -e .
